# Feature/kubosuke/binary search