# Overview

currently fetching the number of last 1min requests sequentially,
this PR is for enhance performance by introducing binary search.

time complexity: O(N) => O(logN)

# Test

100rps up to 30k

## Before (O(N))

```
❯ ab -c 100 -n 15000 http://localhost:8080/
This is ApacheBench, Version 2.3 <$Revision: 1903618 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 1500 requests
Completed 3000 requests
Completed 4500 requests
Completed 6000 requests
Completed 7500 requests
Completed 9000 requests
Completed 10500 requests
Completed 12000 requests
Completed 13500 requests
Completed 15000 requests
Finished 15000 requests


Server Software:
Server Hostname:        localhost
Server Port:            8080

Document Path:          /
Document Length:        3 bytes

Concurrency Level:      100
Time taken for tests:   4.662 seconds
Complete requests:      15000
Failed requests:        14909
   (Connect: 0, Receive: 0, Length: 14909, Exceptions: 0)
Total transferred:      1818846 bytes
HTML transferred:       78846 bytes
Requests per second:    3217.70 [#/sec] (mean)
Time per request:       31.078 [ms] (mean)
Time per request:       0.311 [ms] (mean, across all concurrent requests)
Transfer rate:          381.02 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    1  22.6      0    1055
Processing:     0   24  86.4     14    1096
Waiting:        0   22  81.0     13    1091
Total:          0   25  89.2     14    1096

Percentage of the requests served within a certain time (ms)
  50%     14
  66%     18
  75%     21
  80%     23
  90%     28
  95%     35
  98%     45
  99%    343
 100%   1096 (longest request)
```

## After (O(logN))

```
❯ ab -c 100 -n 15000 http://localhost:8080/
This is ApacheBench, Version 2.3 <$Revision: 1903618 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 1500 requests
Completed 3000 requests
Completed 4500 requests
Completed 6000 requests
Completed 7500 requests
Completed 9000 requests
Completed 10500 requests
Completed 12000 requests
Completed 13500 requests
Completed 15000 requests
Finished 15000 requests


Server Software:
Server Hostname:        localhost
Server Port:            8080

Document Path:          /
Document Length:        2 bytes

Concurrency Level:      100
Time taken for tests:   3.181 seconds
Complete requests:      15000
Failed requests:        14989
   (Connect: 0, Receive: 0, Length: 14989, Exceptions: 0)
Total transferred:      1818863 bytes
HTML transferred:       78863 bytes
Requests per second:    4715.81 [#/sec] (mean)
Time per request:       21.205 [ms] (mean)
Time per request:       0.212 [ms] (mean, across all concurrent requests)
Transfer rate:          558.42 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   6.4      0     552
Processing:     0   21  51.4     14     593
Waiting:        0   20  49.9     13     593
Total:          0   21  51.7     14     594

Percentage of the requests served within a certain time (ms)
  50%     14
  66%     18
  75%     21
  80%     23
  90%     28
  95%     34
  98%     47
  99%    330
 100%    594 (longest request)

```